Refactor the NewsRecyclerAdapter
* Remove the implementation of the NewsItemsClickListener from the
NewsActivity. Creating a new interface to handle the news item click is
an overkill, there is already a View.OnClickListener interface which the
NewsViewHolder can be implement.
* Implement the View.OnclickListener in the NewsViewHolder, to contain
the news list item code within the adapter's view holder and therefore
adhering to the principle of Separation of Concern.
* Implement the LayoutContainer in the NewsViewHolder, invoking
findViewById() can be slow, especially in case of huge view hierarchies,
so Android Extensions tries to minimize findViewById() calls by caching
views in containers.