Rename the NewsInterface to NewsService
This makes it clear on what the file is specifically being used for and
therefore makes it easy to read the code

Renamed the package network to service, naming this network is a bit
confusing as the code contained in it is not related to any networking
but service interfacing. In future this package should contain any files
relating to the service layer of the app.